+++ src/components/common/i18n/utils.ts
@@ -1,8 +1,5 @@
-import type {
-  ICalendarResourceStrings,
-  IDateRangePickerResourceStrings,
-  IResourceStrings,
-} from 'igniteui-i18n-core';
+import type { IResourceStrings } from 'igniteui-i18n-core';
+import { isObject, isString } from '../util.js';
 import type { IgcCalendarResourceStrings } from './EN/calendar.resources.js';
 import type { IgcDateRangePickerResourceStrings } from './EN/date-range-picker.resources.js';
 
@@ -17,9 +14,12 @@ export const validationResourcesKeys = {
   email: 'email_validation_error',
   url: 'url_validation_error',
   disabledDate: 'disabled_date_validation_error',
-};
+} as const;
 
-export const calendarResourcesMap = new Map<string, string | undefined>([
+export const calendarResourcesMap = new Map<
+  keyof IgcCalendarResourceStrings,
+  string | undefined
+>([
   ['selectMonth', 'calendar_select_month'],
   ['selectYear', 'calendar_select_year'],
   ['selectDate', 'calendar_select_date'],
@@ -36,7 +36,10 @@ export const calendarResourcesMap = new Map<string, string | undefined>([
   ['weekLabel', 'i18n/getWeekLabel'],
 ]);
 
-export const dateRangePickerResourcesMap = new Map<string, string | undefined>([
+export const dateRangePickerResourcesMap = new Map<
+  keyof IgcDateRangePickerResourceStrings,
+  string | undefined
+>([
   ['separator', 'date_range_picker_date_separator'],
   ['done', 'date_range_picker_done_button'],
   ['cancel', 'date_range_picker_cancel_button'],
@@ -44,70 +47,96 @@ export const dateRangePickerResourcesMap = new Map<string, string | undefined>([
   ['last30Days', 'date_range_picker_last30Days'],
   ['currentMonth', 'date_range_picker_currentMonth'],
   ['yearToDate', 'date_range_picker_yearToDate'],
-  ...calendarResourcesMap,
+  ...(
+    calendarResourcesMap as Map<
+      keyof IgcDateRangePickerResourceStrings,
+      string | undefined
+    >
+  ).entries(),
 ]);
 
 function isCalendarResource(
-  resource: ICalendarResourceStrings | IgcCalendarResourceStrings
-) {
+  resource: unknown
+): resource is IgcCalendarResourceStrings {
   return (
-    ((resource as ICalendarResourceStrings).calendar_select_month !==
-      undefined ||
-      (resource as IgcCalendarResourceStrings).selectMonth !== undefined) &&
-    !isDateRangePickerResource(resource as IgcDateRangePickerResourceStrings)
+    isObject(resource) &&
+    'selectMonth' in resource &&
+    !isDateRangePickerResource(resource)
   );
 }
 
 function isDateRangePickerResource(
-  resource: IDateRangePickerResourceStrings | IgcDateRangePickerResourceStrings
-) {
-  return (
-    (resource as IDateRangePickerResourceStrings)
-      .date_range_picker_last7Days !== undefined ||
-    (resource as IgcDateRangePickerResourceStrings).last7Days !== undefined
-  );
+  resource: unknown
+): resource is IgcDateRangePickerResourceStrings {
+  return isObject(resource) && 'last7Days' in resource;
 }
 
-export function convertToIgcResource<T extends IResourceStrings>(inObject: T) {
-  const result: any = {};
-  let resourceMap = new Map<string, string | undefined>();
-  if (isCalendarResource(inObject)) {
-    resourceMap = calendarResourcesMap;
-  } else if (isDateRangePickerResource(inObject)) {
-    resourceMap = dateRangePickerResourcesMap;
+function getResourceMap<T>(
+  resource: T
+): Map<string, string | undefined> | undefined {
+  if (isCalendarResource(resource)) {
+    return calendarResourcesMap;
   }
 
-  for (const [key, value] of resourceMap) {
-    if (value) {
-      result[key] = inObject[value as keyof IResourceStrings];
+  if (isDateRangePickerResource(resource)) {
+    return dateRangePickerResourcesMap;
+  }
+
+  return undefined;
+}
+
+function getResourceMapForCore<T extends IResourceStrings>(
+  resource: T
+): Map<string, string | undefined> | undefined {
+  if ('date_range_picker_last7Days' in resource) {
+    return dateRangePickerResourcesMap;
+  }
+
+  if ('calendar_select_month' in resource) {
+    return calendarResourcesMap;
+  }
+
+  return undefined;
+}
+
+export function convertToIgcResource<T extends object>(
+  resource: IResourceStrings
+): T {
+  const result = {} as T;
+  const resourceMap = getResourceMapForCore(resource);
+
+  if (!resourceMap) {
+    return resource as T;
+  }
+
+  for (const [componentKey, coreKey] of resourceMap) {
+    if (coreKey && coreKey in resource) {
+      const coreValue = resource[coreKey as keyof IResourceStrings];
+
+      if (isString(coreValue)) {
+        result[componentKey as keyof T] = coreValue as T[keyof T];
+      }
     }
   }
 
   return result;
 }
 
-export function convertToCoreResource<T>(inObject: T): IResourceStrings {
-  let result = {} as IResourceStrings;
-  let resourceMap: Map<string, string | undefined> | undefined;
-  if (isCalendarResource(inObject as IgcCalendarResourceStrings)) {
-    resourceMap = calendarResourcesMap;
-  } else if (
-    isDateRangePickerResource(inObject as IgcDateRangePickerResourceStrings)
-  ) {
-    resourceMap = dateRangePickerResourcesMap;
-  }
+export function convertToCoreResource<T>(resource: T): IResourceStrings {
+  const result: IResourceStrings = {};
+  const resourceMap = getResourceMap(resource);
 
   if (resourceMap) {
-    for (const [key, value] of resourceMap) {
-      if (value) {
-        result[value as keyof IResourceStrings] = inObject[
-          key as keyof T
-        ] as string;
+    for (const [key, coreKey] of resourceMap) {
+      if (coreKey) {
+        const value = resource[key as keyof T];
+        if (isString(value)) {
+          result[coreKey as keyof IResourceStrings] = value;
+        }
       }
     }
   } else {
-    // If there are no available maps, assume that the input object is of `igniteui-i18n-core` type.
-    result = inObject as IResourceStrings;
+    return resource as IResourceStrings;
   }
 
   return result;