+++ src/components/date-time-input/date-time-input.ts
@@ -23,7 +23,7 @@ import { partMap } from '../common/part-map.js';
 import type { IgcInputComponentEventMap } from '../input/input-base.js';
 import {
   IgcMaskInputBaseComponent,
-  type MaskRange,
+  type MaskSelection,
 } from '../mask-input/mask-input-base.js';
 import IgcValidationContainerComponent from '../validation-container/validation-container.js';
 import {
@@ -253,15 +253,6 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
     }
   }
 
-  @watch('prompt', { waitUntilFirstUpdate: true })
-  protected promptChange(): void {
-    if (!this.prompt) {
-      this.prompt = this.parser.prompt;
-    } else {
-      this.parser.prompt = this.prompt;
-    }
-  }
-
   protected get hasDateParts(): boolean {
     const parts =
       this._inputDateParts ||
@@ -298,11 +289,11 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
   private get targetDatePart(): DatePart | undefined {
     let result: DatePart | undefined;
 
-    if (this.focused) {
+    if (this._focused) {
       const partType = this._inputDateParts.find(
         (p) =>
-          p.start <= this.inputSelection.start &&
-          this.inputSelection.start <= p.end &&
+          p.start <= this._inputSelection.start &&
+          this._inputSelection.start <= p.end &&
           p.type !== DateParts.Literal
       )?.type as string as DatePart;
 
@@ -356,7 +347,7 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
       return;
     }
 
-    const { start, end } = this.inputSelection;
+    const { start, end } = this._inputSelection;
     const newValue = this.trySpinValue(targetPart, delta);
     this.value = newValue;
     this.updateComplete.then(() => this.input.setSelectionRange(start, end));
@@ -370,7 +361,7 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
       return;
     }
 
-    const { start, end } = this.inputSelection;
+    const { start, end } = this._inputSelection;
     const newValue = this.trySpinValue(targetPart, delta, true);
     this.value = newValue;
     this.updateComplete.then(() => this.input.setSelectionRange(start, end));
@@ -378,25 +369,25 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
 
   /** Clears the input element of user input. */
   public clear(): void {
-    this.maskedValue = '';
+    this._maskedValue = '';
     this.value = null;
   }
 
   protected setToday() {
     this.value = new Date();
-    this.handleInput();
+    this._fireInputEvent();
   }
 
   protected updateMask() {
-    if (this.focused) {
-      this.maskedValue = this.getMaskedValue();
+    if (this._focused) {
+      this._maskedValue = this.getMaskedValue();
     } else {
       if (!DateTimeUtil.isValidDate(this.value)) {
-        this.maskedValue = '';
+        this._maskedValue = '';
         return;
       }
 
-      this.maskedValue = DateTimeUtil.formatDisplayDate(
+      this._maskedValue = DateTimeUtil.formatDisplayDate(
         this.value,
         this.locale,
         this.displayFormat,
@@ -405,41 +396,40 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
     }
   }
 
-  protected override handleInput() {
+  private _fireInputEvent(): void {
     this._setTouchedState();
     this.emitEvent('igcInput', { detail: this.value?.toString() });
   }
 
   protected handleDragLeave() {
-    if (!this.focused) {
+    if (!this._focused) {
       this.updateMask();
     }
   }
 
   protected handleDragEnter() {
-    if (!this.focused) {
-      this.maskedValue = this.getMaskedValue();
+    if (!this._focused) {
+      this._maskedValue = this.getMaskedValue();
     }
   }
 
-  protected async updateInput(string: string, range: MaskRange) {
-    const { value, end } = this.parser.replace(
-      this.maskedValue,
-      string,
-      range.start,
-      range.end
-    );
+  protected async _updateInput(
+    text: string,
+    { start, end }: MaskSelection
+  ): Promise<void> {
+    const result = this._parser.replace(this._maskedValue, text, start, end);
 
-    this.maskedValue = value;
+    this._maskedValue = result.value;
 
     this.updateValue();
     this.requestUpdate();
 
-    if (range.start !== this.inputFormat.length) {
-      this.handleInput();
+    if (start !== this.inputFormat.length) {
+      this._fireInputEvent();
     }
+
     await this.updateComplete;
-    this.input.setSelectionRange(end, end);
+    this.input.setSelectionRange(result.end, result.end);
   }
 
   private trySpinValue(
@@ -488,7 +478,7 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
           (dp) => dp.type === DateParts.AmPm
         );
         if (formatPart !== undefined) {
-          amPmFromMask = this.maskedValue.substring(
+          amPmFromMask = this._maskedValue.substring(
             formatPart!.start,
             formatPart!.end
           );
@@ -502,16 +492,16 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
 
   @eventOptions({ passive: false })
   private async onWheel(event: WheelEvent) {
-    if (!this.focused || this.readOnly) {
+    if (!this._focused || this.readOnly) {
       return;
     }
 
     event.preventDefault();
     event.stopPropagation();
 
-    const { start, end } = this.inputSelection;
+    const { start, end } = this._inputSelection;
     event.deltaY > 0 ? this.stepDown() : this.stepUp();
-    this.handleInput();
+    this._fireInputEvent();
 
     await this.updateComplete;
     this.setSelectionRange(start, end);
@@ -544,12 +534,7 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
       '0'
     );
 
-    this._mask = newMask.includes('tt')
-      ? newMask.replace(/tt/g, 'LL')
-      : newMask;
-
-    this.parser.mask = this._mask;
-    this.parser.prompt = this.prompt;
+    this.mask = newMask.includes('tt') ? newMask.replace(/tt/g, 'LL') : newMask;
 
     if (!this.placeholder || oldFormat === this.placeholder) {
       this.placeholder = value;
@@ -563,7 +548,7 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
   }
 
   private getMaskedValue(): string {
-    let mask = this.emptyMask;
+    let mask = this._parser.emptyMask;
 
     if (DateTimeUtil.isValidDate(this.value)) {
       for (const part of this._inputDateParts) {
@@ -577,7 +562,7 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
           this.value
         );
 
-        mask = this.parser.replace(
+        mask = this._parser.replace(
           mask,
           targetValue,
           part.start,
@@ -587,16 +572,20 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
       return mask;
     }
 
-    return this.maskedValue === '' ? mask : this.maskedValue;
+    if (this.readOnly) {
+      return '';
+    }
+
+    return this._maskedValue === '' ? mask : this._maskedValue;
   }
 
   private isComplete(): boolean {
-    return !this.maskedValue.includes(this.prompt);
+    return !this._maskedValue.includes(this.prompt);
   }
 
   private updateValue(): void {
     if (this.isComplete()) {
-      const parsedDate = this.parseDate(this.maskedValue);
+      const parsedDate = this.parseDate(this._maskedValue);
       this.value = DateTimeUtil.isValidDate(parsedDate) ? parsedDate : null;
     } else {
       this.value = null;
@@ -608,7 +597,7 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
   }
 
   private getNewPosition(value: string, direction = 0): number {
-    const cursorPos = this.selection.start;
+    const cursorPos = this._maskSelection.start;
 
     if (!direction) {
       // Last literal before the current cursor position or start of input value
@@ -626,7 +615,7 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
   }
 
   protected async handleFocus() {
-    this.focused = true;
+    this._focused = true;
 
     if (this.readOnly) {
       return;
@@ -636,7 +625,7 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
     const areFormatsDifferent = this.displayFormat !== this.inputFormat;
 
     if (!this.value) {
-      this.maskedValue = this.emptyMask;
+      this._maskedValue = this._parser.emptyMask;
       await this.updateComplete;
       this.select();
     } else if (areFormatsDifferent) {
@@ -645,18 +634,18 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
   }
 
   protected handleBlur() {
-    const isEmptyMask = this.maskedValue === this.emptyMask;
+    const isEmptyMask = this._maskedValue === this._parser.emptyMask;
 
-    this.focused = false;
+    this._focused = false;
 
     if (!(this.isComplete() || isEmptyMask)) {
-      const parse = this.parseDate(this.maskedValue);
+      const parse = this.parseDate(this._maskedValue);
 
       if (parse) {
         this.value = parse;
       } else {
         this.value = null;
-        this.maskedValue = '';
+        this._maskedValue = '';
       }
     } else {
       this.updateMask();
@@ -678,9 +667,9 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
 
   protected async keyboardSpin(direction: 'up' | 'down') {
     direction === 'up' ? this.stepUp() : this.stepDown();
-    this.handleInput();
+    this._fireInputEvent();
     await this.updateComplete;
-    this.setSelectionRange(this.selection.start, this.selection.end);
+    this.setSelectionRange(this._maskSelection.start, this._maskSelection.end);
   }
 
   protected override renderInput() {
@@ -689,22 +678,22 @@ export default class IgcDateTimeInputComponent extends EventEmitterMixin<
         type="text"
         part=${partMap(this.resolvePartNames('input'))}
         name=${ifDefined(this.name)}
-        .value=${live(this.maskedValue)}
-        .placeholder=${live(this.placeholder || this.emptyMask)}
+        .value=${live(this._maskedValue)}
+        .placeholder=${this.placeholder || this._parser.emptyMask}
         ?readonly=${this.readOnly}
         ?disabled=${this.disabled}
         @blur=${this.handleBlur}
         @focus=${this.handleFocus}
-        @input=${super.handleInput}
+        @input=${this._handleInput}
         @wheel=${this.onWheel}
-        @keydown=${super.handleKeydown}
-        @click=${this.handleClick}
-        @cut=${this.handleCut}
-        @compositionstart=${this.handleCompositionStart}
-        @compositionend=${this.handleCompositionEnd}
+        @keydown=${this._setMaskSelection}
+        @click=${this._handleClick}
+        @cut=${this._setMaskSelection}
+        @compositionstart=${this._handleCompositionStart}
+        @compositionend=${this._handleCompositionEnd}
         @dragenter=${this.handleDragEnter}
         @dragleave=${this.handleDragLeave}
-        @dragstart=${this.handleDragStart}
+        @dragstart=${this._setMaskSelection}
       />
     `;
   }